[SWM-163] Feat : calculate playlist duration and save

+++ src/main/java/com/m9d/sroom/util/DateUtil.java
@@ -11,7 +11,6 @@ import java.util.TimeZone;
 
 /**
  * Date Utility
- *
  */
 @Slf4j
 @Service
@@ -51,10 +50,10 @@ public class DateUtil {
         }
     }
 
-    public Long convertISOToSeconds(String isoTime){
+    public int convertISOToSeconds(String isoTime) {
         Duration duration = Duration.parse(isoTime);
 
-        Long totalSeconds = duration.toSeconds();
+        int totalSeconds = (int) duration.toSeconds();
         return totalSeconds;
     }
 
@@ -75,9 +74,9 @@ public class DateUtil {
     /**
      * 2개 시간의 차이를 초 단위로 구한다.
      *
-     * @param   startTime   시작 시간
-     * @param   endTime     종료 시간
-     * @return  long        시작 시간과 종료 시간 사이의 초를 구한다.
+     * @param startTime 시작 시간
+     * @param endTime   종료 시간
+     * @return long        시작 시간과 종료 시간 사이의 초를 구한다.
      */
     public int getDiffSeconds(Date startTime, Date endTime) {
         int dueSeconds = (int) Math.ceil((endTime.getTime() - startTime.getTime()) / 1000);
@@ -88,9 +87,9 @@ public class DateUtil {
     /**
      * 2개 시간의 차이를 일 단위로 구한다.
      *
-     * @param   startTime   시작 시간
-     * @param   endTime     종료 시간
-     * @return  long        시작 시간과 종료 시간 사이의 일을 구한다.
+     * @param startTime 시작 시간
+     * @param endTime   종료 시간
+     * @return long        시작 시간과 종료 시간 사이의 일을 구한다.
      */
     public int getDiffDays(Date startTime, Date endTime) {
         log.info("endTime={}, startTime={}", endTime.getTime(), startTime.getTime());
@@ -102,7 +101,7 @@ public class DateUtil {
     /**
      * 시작 시간으로부터 현재 시간까지의 소요시간을 구해, 로그를 출력한다.
      *
-     * @param   startTime   시작 시간
+     * @param startTime 시작 시간
      */
     public void measureTime(String jobName, Date startTime) {
         Date endTime = new Date();
